Subscriptions: cast function results to `(int)` so strict comparisons pass.
This change fixes a bug where subscription checkboxes and UI elements could show/use incorrect values.

Props wpdennis. Fixes #3206.



git-svn-id: https://bbpress.svn.wordpress.org/trunk@6829 9866e705-20ec-0310-96e7-cbb4277adcfb

+++ src/includes/forums/template.php
@@ -2234,7 +2234,7 @@ function bbp_form_forum_subscribed() {
 		} elseif ( bbp_is_forum_edit() || bbp_is_reply_edit() ) {
 
 			// Get current posts author
-			$post_author = bbp_get_global_post_field( 'post_author', 'raw' );
+			$post_author = (int) bbp_get_global_post_field( 'post_author', 'raw' );
 
 			// Post author is not the current user
 			if ( bbp_get_current_user_id() !== $post_author ) {
